+++ gdrive_sensor/server.py
@@ -27,10 +27,12 @@ from gdrive_sensor import START_PAGE_TOKEN, NEXT_PAGE_TOKEN
 
 logger = logging.getLogger(__name__)
 
+
 def reset_backfill_parameters():
     global START_PAGE_TOKEN, NEXT_PAGE_TOKEN
     node.config.gdrive.start_page_token = START_PAGE_TOKEN
     node.config.gdrive.next_page_token = NEXT_PAGE_TOKEN
+
 
 async def backfill_loop():
     global START_PAGE_TOKEN, NEXT_PAGE_TOKEN
