+++ gdrive_sensor/utils/old_functions.py
@@ -1,0 +1,31 @@
+# from . import get_parent_ids
+# from .connection import drive_service, doc_service, sheet_service, slides_service
+# from .types import GoogleWorkspaceApp, docsType, folderType, sheetsType, presentationType
+# from rid_lib.ext import Cache, Effector, Bundle
+# from koi_net.processor.knowledge_object import KnowledgeObject, RID
+# from koi_net.protocol.event import Event, EventType
+# from googleapiclient.errors import HttpError
+# from gdrive_sensor import SENSOR, SHARED_DRIVE_ID
+
+
+# # cache = Cache(f"{SENSOR}/my_cache")
+# # cache = Cache(f"{ROOT}/net/metadata/gdrive_sensor_node_rid_cache")
+# # cache = Cache(f"{SENSOR}/my_cache")
+# # effector = Effector(cache)
+
+
+
+
+
+
+   
+
+
+
+
+
+
+
+
+
+
