+++ sub.py
@@ -1,0 +1,42 @@
+import uuid
+from googleapiclient.errors import HttpError
+from gdrive_sensor.utils.connection import drive_service
+from gdrive_sensor import SHARED_DRIVE_ID, START_PAGE_TOKEN
+from pprint import pprint
+
+def subscribe_to_drive_changes(driveId, start_page_token, host: str = '0.0.0.0'):
+    channel_id = str(uuid.uuid4())  # Generate a unique channel ID
+    channel_address = f'https://{host}/notifications'  # Your webhook URL
+    resource = {
+        'id': channel_id,
+        'type': 'web_hook',
+        'address': channel_address,
+        'params': {
+            'ttl': 3600  # Time-to-live for the channel in seconds
+        }
+    }
+
+    try:
+        # Call the changes.watch method
+        response = drive_service.changes().watch(
+          driveId=driveId, 
+          includeItemsFromAllDrives=True, 
+          supportsAllDrives=True,
+          includeRemoved=True,
+          spaces='drive',
+          pageToken=start_page_token,
+          body=resource
+        ).execute()
+        # print(f"Subscribed to Drive changes with channel ID: {response['id']}")
+        # print()
+        pprint(response)
+        return response['id']
+    except HttpError as error:
+        print(f"An error occurred: {error}")
+
+# start_page_token = fetch_start_page_token(service=drive_service)
+start_page_token = START_PAGE_TOKEN
+channel_id = subscribe_to_drive_changes(driveId=SHARED_DRIVE_ID, start_page_token=start_page_token, host='0.0.0.0')
+print()
+print(start_page_token)
+print(channel_id)
